Add improved json middlewares and rename json_middleware to wrap_json.
The new added middlewares mimics the current well tested ring-json
middlewares that has more flexible and more correct implementation.

The old `wrap_json` is keept just for backward compatibility. Its
usage now should be discouraged. This is because it has the limitation
of imply the request parsing and body encoding and the body encoding
is limited to list and dict (when json spec supports also other values).